Admin accesibility added

+++ routes/User.js
@@ -2,7 +2,6 @@ const express = require('express');
 const router = express.Router();
 const { PrismaClient } = require('@prisma/client')
 
-const bcrypt = require('bcryptjs');
 const checkLogin = require('../middlewares/CheckLogin');
 
 const prisma = new PrismaClient()
@@ -12,33 +11,57 @@ const prisma = new PrismaClient()
 // GET ALL USERS
 
 router.get('/', checkLogin, async (req, res) => {
-    try {
+
+    // get LoggedIn user
+    const loggedInUser = await prisma.user.findUnique({
+        where: {
+            id: req.user.id
+        }
+    })
+
+    // ADMIN can get all users, user will get only himself
+    if (loggedInUser.role === 'ADMIN') {
         const users = await prisma.user.findMany()
         res.status(200).json(users)
-    } catch (err) {
-        res.status(404).json({ message: 'Something went wrong', error: err })
+    } else {
+        res.status(200).json(loggedInUser)
     }
+
 
 })
 
 // GET ONE USER
 
 router.get('/:id', checkLogin, async (req, res) => {
-    try {
-        const user = await prisma.user.findUnique({
-            where: {
-                id: req.params.id
-            }
-        })
-        res.status(200).json(user)
-    } catch (err) {
-        res.status(404).json({ message: 'Something went wrong', error: err })
+
+    // admin can get all users and user can only get himself
+
+    if (req.userData.role === 'ADMIN' || req.userData.id === req.params.id) {
+        try {
+            const user = await prisma.user.findUnique({
+                where: {
+                    id: req.params.id
+                }
+            })
+            res.status(200).json(user)
+        } catch (err) {
+            res.status(404).json({ message: 'Something went wrong', error: err })
+        }
+    } else {
+        res.status(401).json({ message: 'Unauthorized' })
     }
+
 })
 
 // CREATE USER
 
 router.post('/', checkLogin, async (req, res) => {
+
+    // only admin can create user
+    if (req.user.role !== 'ADMIN') {
+        return res.status(401).json({ message: 'You are not authorized' })
+    }
+
     try {
         const user = await prisma.user.create({
             data: {
@@ -58,32 +81,62 @@ router.post('/', checkLogin, async (req, res) => {
 // UPDATE USER
 
 router.put('/:id', checkLogin, async (req, res) => {
-    const updatedUser = await prisma.user.update({
+
+    // get logged in user
+    const loggedInUser = await prisma.user.findUnique({
         where: {
-            id: req.params.id
-        },
-        data: {
-            name: req.body.name,
-            email: req.body.email,
-            username: req.body.username,
-            role: req.body.role,
-            password: req.body.password
+            id: req.user.id
         }
     })
 
-    res.status(200).json(updatedUser)
+    // check if user is admin or if user is updating their own profile
+    if (loggedInUser.role === 'ADMIN' || loggedInUser.id === req.params.id) {
+        try {
+            const user = await prisma.user.update({
+                where: {
+                    id: req.params.id
+                },
+                data: {
+                    name: req.body.name,
+                    email: req.body.email,
+                    username: req.body.username,
+                    role: req.body.role,
+                    password: req.body.password
+                }
+            })
+            res.status(200).json(user)
+        } catch (err) {
+            res.status(400).json({ message: err })
+        }
+    } else {
+        res.status(401).json({ message: 'You are not authorized to update this user' })
+    }
+
 })
 
 // DELETE USER
 
 router.delete('/:id', checkLogin, async (req, res) => {
-    const deletedUser = await prisma.user.delete({
+
+    // get logged in user 
+    const loggedInUser = await prisma.user.findUnique({
         where: {
-            id: req.params.id
+            id: req.user.id
         }
     })
 
-    res.status(200).json(deletedUser)
+    // check if logged in user is admin
+    if (loggedInUser.role === 'ADMIN') {
+        const deletedUser = await prisma.user.delete({
+            where: {
+                id: req.params.id
+            }
+        })
+        res.status(200).json(deletedUser)
+    } else {
+        res.status(401).json({ message: 'You are not authorized to delete this user' })
+    }
+
 })
 
 module.exports = router
